backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp==3.9.1
aiofiles==23.2.1
playwright==1.41.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
httpx==0.26.0
python-dotenv==1.0.0
replicate==0.22.0
